# docs/openapi-notices.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "AML Service - Notices & Reports API",
		"description": "API endpoints for SAT Notices and Enhanced Reports",
		"version": "2.0.0"
	},
	"servers": [
		{
			"url": "/api/v1",
			"description": "API v1"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
...
			"ReportAggregation": {
				"type": "object",
				"properties": {
					"alerts": {
						"type": "object",
						"properties": {
							"total": { "type": "integer" },
							"bySeverity": {
								"type": "object",
								"additionalProperties": { "type": "integer" }
							},
							"byStatus": {
								"type": "object",
								"additionalProperties": { "type": "integer" }
							},
							"byRule": {
								"type": "array",
								"items": {
									"type": "object",
									"properties": {
										"ruleId": { "type": "string" },
										"ruleName": { "type": "string" },
										"count": { "type": "integer" }
									}
								}
							},
							"byMonth": {
								"type": "array",
								"items": {
									"type": "object",
									"properties": {
										"month": { "type": "string" },
										"count": { "type": "integer" }
									}
								}
							},
							"avgResolutionDays": { "type": "number" },
							"overdueCount": { "type": "integer" }
						}
					},
					"transactions": {
						"type": "object",
						"properties": {
							"total": { "type": "integer" },
							"totalAmount": { "type": "number" },
							"avgAmount": { "type": "number" },
							"byOperationType": { "type": "object" },
							"byVehicleType": { "type": "object" },
							"byCurrency": { "type": "object" },
							"byMonth": { "type": "array" },
							"topClients": { "type": "array" }
						}
					},
					"clients": {
						"type": "object",
						"properties": {
							"total": { "type": "integer" },
							"byPersonType": { "type": "object" },
							"byCountry": { "type": "object" },
							"withAlerts": { "type": "integer" },
							"newInPeriod": { "type": "integer" }
						}
					},
					"riskIndicators": {
						"type": "object",
						"properties": {
							"highRiskClients": { "type": "integer" },
							"criticalAlerts": { "type": "integer" },
							"overdueSubmissions": { "type": "integer" },
							"complianceScore": { "type": "integer" }
						}
					},
					"comparison": {
						"type": "object",
						"nullable": true,
						"properties": {
							"alertsChange": { "type": "number" },
							"transactionsChange": { "type": "number" },
							"amountChange": { "type": "number" },
							"clientsChange": { "type": "number" }
						}
					}
				}
			},
...
		}
	},
	"security": [{ "bearerAuth": [] }],
	"paths": {
...
		"/reports/aggregate/summary": {
			"get": {
				"summary": "Get executive summary aggregation",
				"tags": ["Reports"],
				"parameters": [
					{
						"name": "periodStart",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "periodEnd",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "comparisonPeriodStart",
						"in": "query",
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "comparisonPeriodEnd",
						"in": "query",
						"schema": { "type": "string", "format": "date-time" }
					},
					{ "name": "clientId", "in": "query", "schema": { "type": "string" } }
				],
				"responses": {
					"200": {
						"description": "Aggregation data",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ReportAggregation" }
							}
						}
					}
				}
			}
		},
		"/reports/aggregate/alerts": {
			"get": {
				"summary": "Get alert metrics aggregation",
				"tags": ["Reports"],
				"parameters": [
					{
						"name": "periodStart",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "periodEnd",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{ "name": "clientId", "in": "query", "schema": { "type": "string" } }
				],
				"responses": {
					"200": { "description": "Alert aggregation data" }
				}
			}
		},
		"/reports/aggregate/transactions": {
			"get": {
				"summary": "Get transaction metrics aggregation",
				"tags": ["Reports"],
				"parameters": [
					{
						"name": "periodStart",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "periodEnd",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{ "name": "clientId", "in": "query", "schema": { "type": "string" } }
				],
				"responses": {
					"200": { "description": "Transaction aggregation data" }
				}
			}
		},
		"/reports/aggregate/clients": {
			"get": {
				"summary": "Get client metrics aggregation",
				"tags": ["Reports"],
				"parameters": [
					{
						"name": "periodStart",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "periodEnd",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{ "name": "clientId", "in": "query", "schema": { "type": "string" } }
				],
				"responses": {
					"200": { "description": "Client aggregation data" }
				}
			}
		},
		"/reports/{id}/aggregation": {
			"get": {
				"summary": "Get aggregation data for a specific report",
				"tags": ["Reports"],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Report aggregation data",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ReportAggregation" }
							}
						}
					}
				}
			}
		}
	}
}